+++ noodles-bcf/src/reader/record.rs
@@ -260,19 +260,19 @@ pub(crate) mod tests {
 
     #[test]
     fn test_read_record() -> Result<(), Box<dyn std::error::Error>> {
-        use noodles_vcf::record::{
-            genotypes::{
-                genotype::{
-                    field::{Key as GenotypeFieldKey, Value as GenotypeFieldValue},
-                    Field as GenotypeField,
+        use noodles_vcf::{
+            header::info::Key as InfoFieldKey,
+            record::{
+                genotypes::{
+                    genotype::{
+                        field::{Key as GenotypeFieldKey, Value as GenotypeFieldValue},
+                        Field as GenotypeField,
+                    },
+                    Genotype, Keys,
                 },
-                Genotype, Keys,
+                info::{field::Value as InfoFieldValue, Field as InfoField},
+                Filters as VcfFilters, Genotypes as VcfGenotypes, Ids, Info as VcfInfo, Position,
             },
-            info::{
-                field::{Key as InfoFieldKey, Value as InfoFieldValue},
-                Field as InfoField,
-            },
-            Filters as VcfFilters, Genotypes as VcfGenotypes, Ids, Info as VcfInfo, Position,
         };
 
         use crate::header::StringMaps;